python-services/requirements.txt
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
PyPDF2==3.0.1
